requirements.txt
accelerate==1.0.1
anthropic==0.49.0
azure_storage==0.37.0
circuitsvis==1.43.3
datasets==3.1.0
flash_attn==2.7.4.post1
httpx==0.28.1
jaxtyping==0.3.1
loguru==0.7.3
matplotlib==3.10.1
numpy==2.2.4
openai==1.72.0
optuna==4.1.0
pandas==2.2.3
peft==0.14.0
plotly==5.24.1
PyYAML==6.0.2
PyYAML==6.0.2
sentence_transformers==3.4.1
torch==2.4.0+cu124
tqdm==4.67.1
transformer_lens==2.15.0
transformers==4.48.2
trl==0.15.2
